+++ src/hash/MD5/Md5.java
@@ -1,9 +1,11 @@
+package MD5;
+
 import java.nio.charset.Charset;
 import java.nio.charset.StandardCharsets;
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 
-public class Sha256 {
+public class Md5 {
 
     private static final Charset UTF_8 = StandardCharsets.UTF_8;
     private static final String OUTPUT_FORMAT = "%-20s:%s";
@@ -15,11 +17,12 @@ public class Sha256 {
     private static byte[] digest(byte[] input) {
         MessageDigest md;
         try {
-            md = MessageDigest.getInstance("SHA256");
+            md = MessageDigest.getInstance("MD5");
         } catch (NoSuchAlgorithmException e) {
             throw new IllegalArgumentException(e);
         }
         byte[] result = md.digest(input);
+        System.out.println(result);
         return result;
     }
 
@@ -35,6 +38,7 @@ public class Sha256 {
         return sb.toString();
     }
 
+
     /*
      * main caller, currently just outputs stuff in the console
      */
@@ -48,10 +52,10 @@ public class Sha256 {
 
         // hash props
 
-        byte[] sha256InBytes = Sha256.digest(message.getBytes(UTF_8));
-        System.out.println(String.format(OUTPUT_FORMAT, "SHA256 (hex) ", bytesToHex(sha256InBytes)));
+        byte[] md5InBytes = Md5.digest(message.getBytes(UTF_8));
+        System.out.println(String.format(OUTPUT_FORMAT, "MD5 (hex) ", bytesToHex(md5InBytes)));
         // fixed length, 16 bytes, 128 bits.
-        System.out.println(String.format(OUTPUT_FORMAT, "SHA256 (length)", sha256InBytes.length));
+        System.out.println(String.format(OUTPUT_FORMAT, "MD5 (length)", md5InBytes.length));
 
     }
