+++ src/hash/SHA256/Sha256.java
@@ -1,9 +1,11 @@
+package SHA256;
+
 import java.nio.charset.Charset;
 import java.nio.charset.StandardCharsets;
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 
-public class Md5 {
+public class Sha256 {
 
     private static final Charset UTF_8 = StandardCharsets.UTF_8;
     private static final String OUTPUT_FORMAT = "%-20s:%s";
@@ -15,12 +17,11 @@ public class Md5 {
     private static byte[] digest(byte[] input) {
         MessageDigest md;
         try {
-            md = MessageDigest.getInstance("MD5");
+            md = MessageDigest.getInstance("SHA256");
         } catch (NoSuchAlgorithmException e) {
             throw new IllegalArgumentException(e);
         }
         byte[] result = md.digest(input);
-        System.out.println(result);
         return result;
     }
 
@@ -36,7 +37,6 @@ public class Md5 {
         return sb.toString();
     }
 
-
     /*
      * main caller, currently just outputs stuff in the console
      */
@@ -50,10 +50,10 @@ public class Md5 {
 
         // hash props
 
-        byte[] md5InBytes = Md5.digest(message.getBytes(UTF_8));
-        System.out.println(String.format(OUTPUT_FORMAT, "MD5 (hex) ", bytesToHex(md5InBytes)));
+        byte[] sha256InBytes = Sha256.digest(message.getBytes(UTF_8));
+        System.out.println(String.format(OUTPUT_FORMAT, "SHA256 (hex) ", bytesToHex(sha256InBytes)));
         // fixed length, 16 bytes, 128 bits.
-        System.out.println(String.format(OUTPUT_FORMAT, "MD5 (length)", md5InBytes.length));
+        System.out.println(String.format(OUTPUT_FORMAT, "SHA256 (length)", sha256InBytes.length));
 
     }
